:sparkles: Add LRU to main()
